# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления внутренней видео документацией Vilib.",
        "title": "Vilib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/accounts/{accountId}/roles": {
            "get": {
                "description": "Возвращает список всех ролей аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account_roles"
                ],
                "summary": "Список ролей аккаунта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AccountRole"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую роль для аккаунта с указанными правами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account_roles"
                ],
                "summary": "Создание роли аккаунта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для создания роли",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{accountId}/user-groups": {
            "get": {
                "description": "Возвращает список групп пользователей аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_groups"
                ],
                "summary": "Список групп пользователей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllUserGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую группу пользователей и добавляет указанных пользователей в неё",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_groups"
                ],
                "summary": "Создание группы пользователей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для создания группы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{accountId}/user-groups/roles": {
            "get": {
                "description": "Возвращает список всех ролей групп для аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group_roles"
                ],
                "summary": "Список ролей групп",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.GroupRole"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую роль для группы пользователей с указанными правами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group_roles"
                ],
                "summary": "Создание роли группы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для создания роли группы",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGroupRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGroupRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{accountId}/user-groups/{groupId}/members": {
            "post": {
                "description": "Добавляет указанных пользователей в существующую группу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_groups"
                ],
                "summary": "Добавление участников в группу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID аккаунта",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID группы пользователей",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса для добавления участников",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddGroupMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AddGroupMemberResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.PermissionMask": {
            "type": "integer",
            "format": "int64",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "DefaultPermissionMask"
            ]
        },
        "dto.AccountRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
        "dto.AddGroupMemberRequest": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.AddGroupMemberResponse": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GroupMember"
                    }
                }
            }
        },
        "dto.CreateAccountRoleRequest": {
            "type": "object",
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
        "dto.CreateAccountRoleResponse": {
            "type": "object",
            "properties": {
                "account_role": {
                    "$ref": "#/definitions/dto.AccountRole"
                }
            }
        },
        "dto.CreateGroupRoleRequest": {
            "type": "object",
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permission_mask": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
        "dto.CreateGroupRoleResponse": {
            "type": "object",
            "properties": {
                "role": {
                    "$ref": "#/definitions/dto.GroupRole"
                }
            }
        },
        "dto.CreateUserGroupRequest": {
            "type": "object",
            "required": [
                "name",
                "users"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 64
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.CreateUserGroupResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GroupMember"
                    }
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GroupMember": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.GroupRole": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permission_mask": {
                    "$ref": "#/definitions/domain.PermissionMask"
                }
            }
        },
...
        "dto.GetAllUserGroupsResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserGroup"
                    }
                }
            }
        },
        "dto.UserGroup": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}